Add x86 support to the minidump backend
